Rename elements in the auth feature
