+++ src/auth/components/AuthenticationForm/index.test.tsx
@@ -6,7 +6,7 @@ import { MainAuthenticationForm } from "./index"
 describe("MainAuthenticationForm", () => {
   // Renders the form with default values and no errors
   it("should render the form with default values and no errors", () => {
-    // Render the MainAuthenticationForm component
+    // Render the AuthenticationForm component
     render(<MainAuthenticationForm />)
 
     // Assert that the form is rendered with the default values
@@ -17,7 +17,7 @@ describe("MainAuthenticationForm", () => {
   // Submits the form with valid data and no errors
   it("should submit the form with valid data and no errors", () => {
     // Test code using React Testing Library and hook testing library
-    // Render the MainAuthenticationForm component
+    // Render the AuthenticationForm component
     // Fill in the form with valid data
     // Submit the form
     // Assert that the form is submitted successfully
@@ -27,7 +27,7 @@ describe("MainAuthenticationForm", () => {
   // Toggles between login and register forms
   it("should toggle between login and register forms", () => {
     // Test code using React Testing Library and hook testing library
-    // Render the MainAuthenticationForm component
+    // Render the AuthenticationForm component
     // Assert that the login form is initially displayed
     // Toggle to the register form
     // Assert that the register form is displayed
@@ -38,7 +38,7 @@ describe("MainAuthenticationForm", () => {
   // Submits the form with invalid data and displays errors
   it("should submit the form with invalid data and display errors", () => {
     // Test code using React Testing Library and hook testing library
-    // Render the MainAuthenticationForm component
+    // Render the AuthenticationForm component
     // Fill in the form with invalid data
     // Submit the form
     // Assert that the form is not submitted
@@ -48,7 +48,7 @@ describe("MainAuthenticationForm", () => {
   // Submits the form with an email that is already in use and displays an error
   it("should submit the form with an email that is already in use and display an error", () => {
     // Test code using React Testing Library and hook testing library
-    // Render the MainAuthenticationForm component
+    // Render the AuthenticationForm component
     // Fill in the form with an email that is already in use
     // Submit the form
     // Assert that an error message is displayed for the email field
@@ -57,7 +57,7 @@ describe("MainAuthenticationForm", () => {
   // Submits the login form with incorrect email or password and displays an error
   it("should submit the login form with incorrect email or password and display an error", () => {
     // Test code using React Testing Library and hook testing library
-    // Render the MainAuthenticationForm component
+    // Render the AuthenticationForm component
     // Fill in the login form with incorrect email or password
     // Submit the form
     // Assert that an error message is displayed for the login form
