+++ src/pages/login.page.tsx
@@ -1,4 +1,4 @@
-import { MainAuthenticationForm } from "@/auth/components/MainAuthenticationForm"
+import { MainAuthenticationForm } from "src/auth/components/AuthenticationForm"
 import { BlitzPage, Routes } from "@blitzjs/next"
 import Layout from "@/core/layouts/Layout"
